+++ Changelog.md
@@ -15,6 +15,7 @@
 ### Removal
 
 * Remove unused `Webdriver.web_console_error`
+* Remove unused `Webdriver#set_text_to_iframe`
 
 ## 0.1.2 (2020-02-18)